+++ src/components/common/chip/ColorChip.module.scss
@@ -1,6 +1,7 @@
 @import '@/styles/color.scss';
 @import '@/styles/responsive-styles.scss';
 
+// ColorChip
 .colorChipContainer {
   position: relative;
   display: inline-flex;
