+++ src/components/common/chip/ColorChip.tsx
@@ -5,6 +5,7 @@ import { HexColorPicker } from 'react-colorful';
 import Image from 'next/image';
 import styles from './ColorChip.module.scss';
 
+// ColorChip
 // Image
 const COLORS = ['#7ac555', '#760dde', '#ffa500', '#76a5ea', '#e876ea'];
