+++ src/components/common/chip/LabelChip.module.scss
@@ -1,5 +1,6 @@
 @import '@/styles/responsive-styles.scss';
 
+// label
 .labelChip {
   &.columns {
     display: inline-flex;
