+++ src/components/common/chip/NumberChip.module.scss
@@ -1,5 +1,6 @@
 @import '@/styles/color.scss';
 
+// number
 .numberChip {
   display: inline-flex;
   padding: 0.1875rem 0.375rem;
