+++ src/components/common/chip/NumberChip.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import styles from './numberChip.module.scss';
 
+// number
 interface NumberChipProps {
   number: number;
 }
