+++ src/components/common/chip/PlusChip.module.scss
@@ -1,6 +1,7 @@
 @import '@/styles/color.scss';
 @import '@/styles/responsive-styles.scss';
 
+// plus
 .plusChip {
   display: flex;
   justify-content: center;
