+++ src/components/common/chip/PlusChip.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import Image from 'next/image';
 import styles from './PlusChip.module.scss';
 
-// to plusBtn
 export default function PlusChip(): JSX.Element {
   return (
     <div className={`${styles.plusChip}`}>
