+++ src/components/common/input/PasswordInput/PasswordInput.module.scss
@@ -1,5 +1,6 @@
 @import '@/styles/color.scss';
 
+// label
 .label {
   font-size: 16px;
   display: flex;
