+++ src/components/modal/CreateDashboard/CreateDashboard.module.scss
@@ -2,6 +2,7 @@
 @import '@/styles/_mixins.scss';
 @import '@/styles/_responsive-styles.scss';
 
+// create
 .container {
   display: flex;
   flex-direction: column;
