+++ src/components/modal/DeleteTask/DeleteTask.module.scss
@@ -1,5 +1,6 @@
 @import '@/styles/_color.scss';
 
+// delete task
 .container {
   display: flex;
   text-align: center;
@@ -17,7 +18,7 @@
     font-style: normal;
     line-height: normal;
     margin-bottom: 50px;
-    
+
     span {
       font-size: 18px;
       font-weight: 500;
