+++ src/components/modal/Modal.module.scss
@@ -1,5 +1,6 @@
 @import '@/styles/_variables.scss';
 
+// modal
 .container {
   z-index: $z-2;
   position: fixed;
