+++ src/components/modal/ModalButton/SubmitButton/ModalSubmitButton.module.scss
@@ -2,6 +2,7 @@
 @import '@/styles/_mixins.scss';
 @import '@/styles/_responsive-styles.scss';
 
+// button
 .container {
   @include flex-center;
   width: 7.5rem;
