+++ src/components/modal/Task/Kebob/KabobMenu.module.scss
@@ -1,7 +1,7 @@
 @import '@/styles/_color.scss';
 @import '@/styles/_variables.scss';
 
-.kabobMenuContainer {
+.kebobMenuContainer {
   position: relative;
 
   .menuOptions {
