--- src/components/modal/Task/mock.json
@@ -1,18 +1,0 @@
-{
-  "id": 4850,
-  "title": "a",
-  "description": "a",
-  "tags": ["ㅁ"],
-  "dueDate": "2024-04-18 12:00",
-  "assignee": {
-    "id": 1520,
-    "nickname": "밍",
-    "profileImageUrl": "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/3-7_1520_1712645279291.png"
-  },
-  "imageUrl": "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/task_image/3-7_19352_1713702961720.jpeg",
-  "teamId": "3-7",
-  "columnId": 19352,
-  "dashboardId": 5753,
-  "createdAt": "2024-04-12T14:55:14.923Z",
-  "updatedAt": "2024-04-21T21:36:04.538Z"
-}
