+++ src/components/modal/index.tsx
@@ -1,6 +1,7 @@
 import React, { ReactNode, useRef, useEffect } from 'react';
 import styles from './Modal.module.scss';
 
+// modal
 interface ModalProps {
   children: ReactNode;
   isOpen: boolean;
