+++ src/components/modal/modalInput/ModalInput.module.scss
@@ -1,4 +1,6 @@
 @import '@/styles/color.scss';
+
+// modalInput
 .container {
   position: relative;
   width: 100%;
