+++ src/components/modal/modalInput/commentInput/CommentInput.module.scss
@@ -1,6 +1,7 @@
 @import '@/styles/color.scss';
 @import '@/styles/responsive-styles.scss';
 
+// common
 .commentForm {
   display: flex;
   position: relative;
