+++ src/components/modal/modalInput/commentInput/index.tsx
@@ -1,5 +1,3 @@
-/* eslint-disable no-console */
-
 'use client';
 
 import React, { ChangeEvent, useEffect, useState } from 'react';
