+++ src/components/modal/modalInput/deadlineInput/DeadlineInput.module.scss
@@ -1,4 +1,5 @@
 @import '@/styles/color.scss';
+// deadLine
 .container {
   .deadline {
     font-size: 1.125rem;
