+++ src/components/modal/modalInput/tagInput/TagInput.module.scss
@@ -1,4 +1,5 @@
 @import '@/styles/color.scss';
+// tag input
 .container {
   .tagText {
     font-size: 1.125rem;
