+++ src/components/modal/modalInput/tagInput/index.tsx
@@ -4,6 +4,7 @@ import React, { ChangeEvent, useState, KeyboardEvent } from 'react';
 import LabelChip from '@/components/common/Chip/LabelChip';
 import styles from './TagInput.module.scss';
 
+// tag input
 export default function TagInput() {
   const [tag, setTag] = useState('');
   const [tagList, setTagList] = useState<string[]>([]);
