+++ src/components/modal/modalInvite/ModalInvite.module.scss
@@ -1,3 +1,5 @@
+// invite
+
 .inviteIcon {
   fill: #ffffff;
 }
